refactor(cpp): extract helpers and builder
